Moved project client and model objects to unique package
